EDIT dl all only students, name as answer author
Changes to the  "Download All" attachments button in the "Participation"
tab of an assignment.

Instructor submitted answers were being included as part of this.
However, it seems that the "Participation" tab only lists student
answers. It's probably better to be consistent with the "Participation"
tab and also only include student answer attachments. As such, I've
modified the API so that it's downloading student answer attachments
only.

I didn't know that instructors can submit answers on a student's behalf.
This caused an interesting issue where the file is linked to the
instructor but the answer is linked to the student. Since the file name
comes from the file user, the file gets the instructor's name.  This is
different from the UI, where the answer user is used. It makes more
sense to use the answer user like the UI, so I've switched it to use the
answer user in the filename.

Tests were updated. Interestingly, the tests actually assumes usage of
the answer user, so we didn't have to do much there.