EDIT elo library now local
Original library is no longer maintained and has issues with setup.py
calling use_2to3 which was removed in setuptools v58. Since we just need
elo.py, I've just copied it locally so we can skip using it as a
dependency to be installed.

Minor modifications were made:
- metaclass removed, python 3 has new syntax and it doesn't seem
necessary
- long datatype removed, python 3 no longer has long

This fixes #1028