UPGRADE to flask 2.0, use pipenv for pkg control
This commit is just a start, a lot of things still doesn't work. So far,
two libs that were using needs to be replaced:
    1. flask-script
    2. flask-restful

Both are no longer maintained and causes runtime exceptions.
Flask-script tries to import flask._compat which doesn't seem to exist
anymore. Flask-restful's param parser doesn't work with werkzeug 2.1.0
change to request.get_json().

Needed to rename main.py to app.py as since we don't set FLASK_APP env
var, Flask looks for app.py as the default application. This requires us
to change the uwsgi config too to reflect this.

Other changes were made for deprecation reasons.

Pipenv is a package manager like npm/yarn/bundler. It uses pip
underneath and also provides an all-in-one way to work with virtual
environments. I'm hoping this will make it easier to upgrade
dependencies, as previously, I was going through them one by one
manually. This should let us retire from using requirements.txt.

Pipenv defaults to installing to the local venv. But in the case of our
docker containers, we want it to install packages as system wide
packages. This is done by the '--system' param, e.g.:
    pipenv install --system

For local dev though, we can run a 'pipenv install' and a 'pipenv shell'
to install to a local venv and activate the venv.

Note that we're installing pipenv from pip cause the one from apt has a
bug that makes it break when trying to install imsglobal-caliper from
zip file.