FIX user api tests broken by upgrade
Latest Werkzeug removed "as_tuple" which was being used in Flask tests.
Needed to update Flask to at least 2.0.3 to fix it. I don't want to
bring Flask up to the latest until all the upgrade issues have been
fixed.

Changed the webargs error handler to return messages that are at least
displayable by the front end.

Moved compair.util.fields to compair.api.util.fields. I think it'll only
be used in the api, hence the move. Added a "not_empty" validator that
can be reused so we don't have to keep copy & pasting it.

Not sure why, but the lti launch to assignment test failed, the param
parsing wasn't working. So switched it over to webargs to fix it.

Fixed error message thrown by UserCourseStatusListAPI. This was being
caught by the webargs validation which throws a 422 instead of the 400
that the tests expected. So I've rethrown webargs 422 to 400. Also
moved the old error message to the webargs error_message. Just in case,
I've also made the frontend treat 422 errors like 400.

Tests need to reset the database. This is done by calling the CLI
database manager. Unfortunately, Flask-Script isn't working anymore. So
I've migrated the database manager over to Flask's new default CLI. All
of the CLI managers needs to be migrated, but I've only done the
database manager to get the tests up and running again.