FIX tests broken by send_file & query params
send_file's attachment_filename param was renamed to download_name. This
needed to be reflected in the test cases.

It looks like Flask-Restful's RequestParser no longer works with query
params (for some tests, maybe not all). At least, it returns a 400 when
it shouldn't. So I've moved them to use webargs. Note that since webargs
doesn't allow reading the same params from multiple locations (or at
least, not simply), these params are now restricted to query only. In
the tests, some of these queries were being sent in the json body of GET
requests, which broke them. So I've modified them to send those params
in the query_string instead of in data.

Also attempted to migrate away from Flask-Restful's Resource by
replacing it with Flask's MethodView in AnswerUserIdAPI. This seems to
be a drop in replacement, though adding the url to the blueprint is now
a bit more wordy.

This commit was made as a save point before attempting to upgrade to
Flask 2.1.0 to pick up the redirect response change. I think the
redirect response change should fix the currently broken LTI launch
test.