UPGRADE flask to latest 2.1.3
This is the latest version of Flask.

Had to override Flask-Testing's implementation of assertRedirect().
Werkzeug now allows relative urls in redirects but Flask-Testing is
still converting them to absolute urls. So the override basically just
takes the expected location as is and won't try to convert relative into
absolute url.

All backend and frontend tests are passing.