FIX lti launch with assignment not working again
Partially revert a previous fix where we found that Canvas was not using
the entire URL (in particular, the query section) to generate
signatures. I suspect this seems to have been fixed on the Canvas side
and they're now using the entire URL for signature verification. So we
need to partially revert the previous fix.