ADD kaltura migration script to cli tools
We're switching from on-prem hosted Kaltura to cloud Kaltura. When
videos are transferred to the new Kaltura instance, a new entry ID is
generated. We will be given a CSV mapping of old to new entry IDs, and
will have to update our Kaltura data accordingly.

The CSV is expected to be in the following format, where the first row
is a header and can be ignored:

old entry id, new entry id
0_someenid,0_newentid
0_someenid,0_newentid
0_someenid,0_newentid

Assuming that collision between old and new entry IDs are impossible.

Requires that these Kaltura env vars are set to the new Kaltura
environment:

* KALTURA_SERVICE_URL
* KALTURA_PARTNER_ID
* KALTURA_SECRET
* KALTURA_USER_ID
* KALTURA_PLAYER_ID

Usage options, run in app root:

    python manage.py kaltura migrate /path/to/mappingCsv.csv

-d Do a dry run, without actually making any changes to the database:

    python manage.py kaltura migrate -d /path/to/mappingCsv.csv

-n If present, tells the CSV parser not to skip the first row. By
default, we assume the first row is a header row and skip it:

    python manage.py kaltura migrate -n /path/to/mappingCsv.csv

This will migrate all the normal Kaltura integration uploaded videos.
There's a handful of special cases where people inserted Kaltura video
links into answers that this does NOT deal with. We're not sure how to
deal with that on the Kaltura side yet, as those videos were uploaded to
a different section that wasn't meant for ComPAIR.