EDIT kaltura migration for links in answer content
We have 28 kaltura videos that didn't go through ComPAIR's Kaltura
integration. The only thing we have of them is a link to the video, the
link is in this format:

https://learning.video.ubc.ca/media/<video name>/<video entry ID>

This video is also on a different Kaltura repo, it's not in the ComPAIR
Kaltura repo. As a special edge case for this migration, we're going to
copy these videos into the ComPAIR Kaltura repo. We'll also make it look
like the video was uploaded via ComPAIR, the video will show up as an
attachment like all the other file attachments.

Note that the original learning.video.ubc.ca link will no longer work
after a certain point and we won't be fixing the link. The video will
only be available via the attachment.

The same mapping mechanism as the regular kaltura migrate command is
used. This differs from the regular migrate command in that we need to
create new entries in the 'file' and 'kaltura_media' table. The command
is similar too:

Dry run:
python manage.py kaltura links -d mapping.csv

For real:
python manage.py kaltura links mapping.csv

Log file will be in format:
kaltura-links-migration-log-<iso8601 datetime>.log